+++ VVExtensions/Bool+Swifty.swift
