+++ VVExtensions/Device+Swifty.swift
