+++ VVExtensions/Info.plist
