+++ VVExtensions/Int+Swifty.swift
