+++ VVExtensions/Optional+Swifty.swift
