+++ VVExtensions/UIImage+Swifty.swift
