+++ VVExtensionsTests/Bool+SwiftyTests.swift
