+++ VVExtensionsTests/String+SwiftyTest.swift
